Criação de inserção Post

+++ boa-viagem-api/src/main/java/br/com/etechoracio/boa_viagem/controller/GastoController.java
@@ -1,12 +1,16 @@
 package br.com.etechoracio.boa_viagem.controller;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -28,16 +32,46 @@ public class GastoController {
 	
 	
 	@GetMapping("/{id}")
-	public Gasto buscarPorId(@PathVariable Long id) {
+	public ResponseEntity<Gasto> buscarPorId(@PathVariable Long id) {
 		
-		return repository.findById(id).orElse(null);
+		Optional<Gasto> existe = repository.findById(id);
+		if (existe.isPresent()) {
+			return ResponseEntity.ok(existe.get());
+		} 
+			return ResponseEntity.notFound().build();
+		
 	}
 	
 	
 	
 	@DeleteMapping("/{id}")
-	public void deletarPorId(@PathVariable Long id) {
+	public ResponseEntity<Object> deletarPorId(@PathVariable Long id) {
+		boolean existe = repository.existsById(id);
 		
-		repository.deleteById(id);
+		if (existe) {
+			
+			repository.deleteById(id);
+			return ResponseEntity.ok().build();
+			
+		}
+			return ResponseEntity.notFound().build();
+			
+	}
+	
+	@PostMapping
+	public ResponseEntity<Gasto> inserir(@RequestBody Gasto obj) {
+		
+		repository.save(obj);
+		return ResponseEntity.ok(obj);
+		
 	}
 }
+
+
+
+
+
+
+
+
+
